Split storage tests
